Move toggle-button to holocene
